move filtering utilities to akri-discovery-utils
