+++ discovery-utils/Cargo.toml
@@ -13,9 +13,9 @@ tonic = { version = "0.1.0", features = ["tls"] }
 # for run_discovery_server
 akri-shared = { path = "../shared" }
 futures = { version = "0.3.1", package = "futures" }
-
-# for registration_client
 log = "0.4"
+serde = "1.0"
+serde_derive = "1.0"
 tokio = { version = "0.2", features = ["time", "net"] }
 tower = "0.3"
